Sync modular-pipeline test split and private boundary helper
Ports the latest two review rounds from snapshot/modular-pipelines.

- compute_namespace_boundaries becomes _compute_namespace_boundaries. Nothing
  outside the package calls it, so it is private.
- Adapter tests mirror the production package:

    tests/test_inspection_adapter/modular_pipelines/
      conftest.py         shared node, tree builder and index factories
      test_boundaries.py  namespace boundary calculation
      test_index.py       which modular pipelines a dataset belongs to
      test_tree.py        folder hierarchy
      test_view.py        group nodes and boundary edges

  The old test_modular_pipelines.py mixed all four concerns, so a tree failure
  now opens test_tree.py. All 22 tests move across unchanged.
