+++ package/tests/test_inspection_adapter/modular_pipelines/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for modular-pipeline rendering from inspection snapshots."""
